Chore : userId를 프론트에서 넘겨주는 부분을 삭제 (로그인한 userId를 백엔드에서 꺼내주는 로직으로 변경)

+++ src/main/java/com/bookpie/shop/service/BookReviewService.java
@@ -1,24 +1,23 @@
 package com.bookpie.shop.service;
 
+import com.bookpie.shop.config.JwtTokenProvider;
 import com.bookpie.shop.domain.BookReview;
 import com.bookpie.shop.domain.User;
 import com.bookpie.shop.domain.dto.book_review.BookReviewDto;
-import com.bookpie.shop.repository.BookRepository;
 import com.bookpie.shop.repository.BookReviewRepository;
 import com.bookpie.shop.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityNotFoundException;
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
@@ -31,11 +30,13 @@ import java.util.stream.Collectors;
 public class BookReviewService {
     private final BookReviewRepository bookReviewRepository;
     private final UserRepository userRepository;
+    private final JwtTokenProvider jwtTokenProvider;
+    private final HttpServletRequest request;
 
     // 도서 리뷰 작성
-    public BookReviewDto create(BookReviewDto dto) {
+    public BookReviewDto create(BookReviewDto dto, Long userId) {
         // 회원 객체 생성
-        User user = userRepository.findById(dto.getUserId())
+        User user = userRepository.findById(userId)
                 .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
 
 
@@ -51,39 +52,48 @@ public class BookReviewService {
     }
 
     // 도서 리뷰 수정
-    public BookReviewDto update(BookReviewDto dto) {
+    public BookReviewDto update(BookReviewDto dto, Long userId) {
         // 해당 리뷰 생성
         BookReview bookReview = bookReviewRepository.findById(dto.getReviewId())
                 .orElseThrow(() -> new IllegalArgumentException("해당 리뷰는 존재하지 않습니다."));
+
+        if (userId != bookReview.getUser().getId())
+            throw new IllegalArgumentException("도서리뷰 수정 실패! 회원 정보가 일치하지 않습니다.");
 
         // 리뷰 수정
         bookReview.patch(dto);
 
         // 리뷰 수정 후 DB저장
         BookReview createdBookReview = bookReviewRepository.save(bookReview);
-        return BookReviewDto.createDto(createdBookReview, dto.getUserId());
+        return BookReviewDto.createDto(createdBookReview, userId);
     }
 
     // 도서 리뷰 삭제
-    public String delete(Long reviewId) {
+    public String delete(Long reviewId, Long userId) {
         // 해당 리뷰 생성
         BookReview bookReview = bookReviewRepository.findById(reviewId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 리뷰는 존재하지 않습니다."));
+
+        if (bookReview.getUser().getId() != userId)
+            throw new IllegalArgumentException("도서리뷰 삭제 실패! 회원 정보가 일치하지 않습니다.");
+
         bookReviewRepository.delete(bookReview);
         return "해당 리뷰가 삭제되었습니다.";
     }
 
     // 해당 도서의 리뷰 전체 조회
-    public JSONObject getReview(String isbn, String page, String size, String userId) {
+    public JSONObject getReview(String isbn, String page, String size) {
         // 반환할 제이슨 객체
         JSONObject response = new JSONObject();
+        Long userId = 0L;
 
-        Long user_id = 0L;
-        if (userId != null) user_id = Long.parseLong(userId);
-
+        // 로그인 유저가 있으면
+        if (jwtTokenProvider.resolveToken(request) != null) {
+            userId = getCurrentUserId();
+        }
         PageRequest pageRequest = PageRequest.of(Integer.parseInt(page), Integer.parseInt(size), Sort.by("reviewDate").descending());  // 페이징 정보
 
-        Long finalUser_id = user_id;
+        Long finalUser_id = userId;
         // Page<BookReview>객체를 Page<BookReviewDto> 객체로 변환
         Page<BookReviewDto> bookReviewDtoPage = bookReviewRepository.findAllByIsbn(isbn, pageRequest)
                 .map(bookReview -> BookReviewDto.createDto(bookReview, finalUser_id));
@@ -131,38 +141,41 @@ public class BookReviewService {
     }
 
     // 해당 책에 내가 쓴 도서리뷰
-    public BookReviewDto myReview(String isbn, String userId) {
-        Long user_id = 0L;
-        if (userId != null) {
-            user_id = Long.parseLong(userId);
-        } else {
+    public BookReviewDto myReview(String isbn, Long userId) {
+        if (userId == 0) {
             return null;
         }
 
-        User user = userRepository.findById(user_id)
+        User user = userRepository.findById(userId)
                 .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
-        BookReview bookReview = bookReviewRepository.findMyReview(isbn, user_id);
+
+        BookReview bookReview = bookReviewRepository.findMyReview(isbn, userId);
         if (bookReview == null) return null;
 
-        return BookReviewDto.createDto(bookReview, user_id);
+        return BookReviewDto.createDto(bookReview, userId);
     }
+
     // 해당 책에서 베스트 리뷰 2개
-    public List<BookReviewDto> bestReview(String isbn, String userId) {
-        Long user_id = Long.parseLong(userId);
-        User user = userRepository.findById(user_id)
-                .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
+    public List<BookReviewDto> bestReview(String isbn) {
+        Long user_id = 0L;
+        if (jwtTokenProvider.resolveToken(request) != null) {
+            user_id = getCurrentUserId();
+            User user = userRepository.findById(user_id)
+                    .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
+        }
 
         List<BookReview> reviewList = bookReviewRepository.bestReview(isbn);
+        Long finalUser_id = user_id;
         return reviewList.stream()
-                .map(bookReview -> BookReviewDto.createDto(bookReview, user_id))
+                .map(bookReview -> BookReviewDto.createDto(bookReview, finalUser_id))
                 .collect(Collectors.toList());
     }
 
-    public List<JSONObject> myCategory(String userId) {
-        Long user_id = Long.parseLong(userId);
-        User user = userRepository.findById(user_id)
+    public List<JSONObject> myCategory(Long userId) {
+        User user = userRepository.findById(userId)
                 .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
-        List<JSONObject> myCategoryList = bookReviewRepository.myCategory(user_id);
+
+        List<JSONObject> myCategoryList = bookReviewRepository.myCategory(userId);
 
         List<JSONObject> resultObjs = new ArrayList<>();
 
@@ -176,5 +189,10 @@ public class BookReviewService {
         }
 
         return resultObjs;
+    }
+
+    private Long getCurrentUserId(){
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        return user.getId();
     }
 }
